Give the executable's CI build a configuration MSIX will accept
The first run of the new step failed in nine seconds, before compiling
anything: "Improper project configuration: WindowsPackageType is set to None,
but AppxPackage is set to true". The project's ApplicationType is Windows
Store, so AppxPackage defaults to true, and asking for an unpackaged build
means saying so twice.

Also in this commit, from reading the shell back rather than from running it:

  * winrt::throw_last_error() replaces check_hresult(HRESULT_FROM_WIN32(
    GetLastError())) on the four Win32 failure paths. The second does NOT throw
    when the last error happens to be zero, which is exactly the case where
    something has gone wrong and there is nothing to diagnose it with.
  * Main.cpp needs Docking.h for DockedShip; nothing else in its include set
    pulls it in.
  * Window's KEY_COUNT is a std::uint8_t rather than a std::size_t, because it
    is only ever compared against one and the promotion makes that a
    signed/unsigned mismatch.

And the title screen gets the PLACEHOLDER BOX the plan's 2e row asks for. Its
size follows the distance the way LL9's projection will, which is the one thing
about a placeholder that can be right: a Cobra settling at 210 is small and an
Adder at 48 fills the screen, and a fixed-size box would make the two title
screens look identical when the game means them to look nothing alike.

+++ Outpost/Window.h
@@ -107,7 +107,7 @@ private:
   static constexpr std::size_t MAX_QUEUED_KEYS = 16;
 
   /// 6502: KEYLOOK is 65 bytes, one per internal key number, which is also TRANTABLE's extent.
-  static constexpr std::size_t KEY_COUNT = 65;
+  static constexpr std::uint8_t KEY_COUNT = 65;
 
   HWND m_window = nullptr;
   HINSTANCE m_instance = nullptr;
